+++ transtats/config/settings/test.py
@@ -20,4 +20,4 @@ DEBUG = False
 EMAIL_BACKEND = 'django.core.mail.backends.console.EmailBackend'
 
 # Allowed Hosts
-ALLOWED_HOSTS = ['localhost', '127.0.0.1']
+ALLOWED_HOSTS = ['*']