+++ transtats/dashboard/managers/transplatform.py
@@ -22,30 +22,6 @@ class TransPlatformManager(BaseManager):
     Translation Platform Manager
     """
 
-    def get_translation_platform(self):
-        """
-        fetch translation platforms from db, zanata as of now!
-        """
-        platform = {}
-        translation_platform = None
-        '''
-        zanata_platform = TransPlatform(engine_name='zanata', api_url='http://translate.zanata.org',
-                                        server_status='active', subject='public', platform_id=3)
-        self.db_session.add(zanata_platform)
-        self.db_session.commit()
-        '''
-        try:
-            translation_platform = self.db_session.query(TransPlatform).filter_by(subject='fedora').first()
-        except:
-            # log event, passing for now
-            pass
-        if translation_platform:
-            platform['url'] = translation_platform.api_url
-            platform['engine'] = translation_platform.engine_name
-            platform['subject'] = translation_platform.subject
-            platform['state'] = translation_platform.server_status
-        return platform
-
     def get_active_transplatforms(self):
         """
         Fetch slug and api_url for all active transplatforms
@@ -54,8 +30,9 @@ class TransPlatformManager(BaseManager):
         platforms = None
         try:
             platforms = self.db_session.query(TransPlatform.platform_slug, TransPlatform.api_url) \
-                .filter_by(server_status='active').all()
+                .filter_by(server_status=True).all()
         except:
+            self.db_session.rollback()
             # log event, passing for now
             pass
         return tuple(platforms) if platforms else ()